Rename run_all_gui to run_dispatch and remove GUI
