+++ arbeitslog.md
@@ -1,0 +1,7 @@
+# Arbeitslog
+
+## 2024-05-16
+- `run_all_gui.py` in `run_dispatch.py` umbenannt und GUI-Code entfernt.
+- Nur `summarize_day` und `process_month` mit `_log` beibehalten.
+- Importpfade auf `run_dispatch` angepasst.
+- Tests und `pyproject.toml` entsprechend aktualisiert.
