+++ pyproject.toml
@@ -17,12 +17,8 @@ dependencies = [
     "pandas>=2.0",
 ]
 
-[project.optional-dependencies]
-gui = ["PySimpleGUI>=5,<6"]
-
 [project.scripts]
 dispatch = "dispatch.main:main"
-dispatch-gui = "run_all_gui:run_gui"
 
 [tool.setuptools.packages.find]
 where = ["."]
